backend/requirements.txt
mysql-python
sqlalchemy
pycrypto
redis
flask
